flask-server/requirements.txt
flask[async]
flask_cors
google-cloud-texttospeech
openai
python-dotenv